úpravy min max boxu

+++ Source/TauCetiF2/Components/SelectorComponent.cpp
@@ -82,7 +82,7 @@ void USelectorComponent::TickComponent(float DeltaTime, ELevelTick TickType, FAc
 
 
 
-	float cubeSize = UHelpers::CubeMinSize;
+	int32 cubeSize = UHelpers::CubeMinSize;
 
 	ImpactPoint = result.ImpactPoint;
 	ImpactNormal = result.ImpactNormal;
